+++ packages/zp-data/src/index.ts
@@ -1,1 +1,1 @@
-export { default as MongoDBService } from './mongodb'
+export * from './mongodb'