+++ packages/zp/src/declaration.ts
@@ -24,14 +24,29 @@ export interface Params<Q = Query> {
   authenticated?: boolean
 }
 
-export interface ServiceMethods<T = any, D = Partial<T>, P = Params> {
-  find?: (params?: P) => Promise<{ total: number, limit?: number, skip: number, data: T[] }>
-  get?: (id: Id, params?: P) => Promise<T>
-  create?: (data: D, params?: P) => Promise<T | T[]>
-  update?: (id: NullableId, data: D, params?: P) => Promise<T | T[]>
-  patch?: (id: NullableId, data: D, params?: P) => Promise<T | T[]>
-  remove?: (id: NullableId, params?: P) => Promise<T | T[]>
+export interface Paginated<R> {
+  total: number
+  limit?: number
+  skip: number
+  data: R[]
 }
 
-export type Service<T = any, D = Partial<T>, P = Params> = ServiceMethods<T, D, P>
+export interface Service<Result = any, Data = Partial<Result>, ServiceParams extends Params = any> {
+  find? (params?: ServiceParams): Promise<Result[]>
+  find? (params?: ServiceParams & { query: { $paginate: true } }): Promise<Paginated<Result>>
+  find? (params?: ServiceParams): Promise<Paginated<Result> | Result[]>
+
+  create? (data: Data, params?: ServiceParams): Promise<Result>
+  create? (data: Data[], params?: ServiceParams): Promise<Result[]>
+
+  patch? (id: null, data: Data, params?: ServiceParams): Promise<Result[]>
+  patch? (id: Id, data: Data, params?: ServiceParams): Promise<Result>
+
+  remove? (id: null, params?: ServiceParams): Promise<Result[]>
+  remove? (id: Id, params?: ServiceParams): Promise<Result>
+
+  get? (id: Id, params?: ServiceParams): Promise<Result>
+  update? (id: Id, data: Data, params?: ServiceParams): Promise<Result>
+}
+
 export { Application }