+++ playground/hello/src/services/status/status.class.ts
@@ -1,17 +1,19 @@
 
-import { Params } from 'zapnode'
+import { Paginated, Params, Service } from 'zapnode'
 import { ConfigData } from '@/declarations'
 
-class AuthClass {
+class StatusClass implements Service {
   constructor (protected config: ConfigData) {}
 
-  async find (params?: Params) {
-    return {
+  async find (params?: Params): Promise<any[]>
+  async find (params?: Params & { query: { $paginate: true } }): Promise<Paginated<any>>
+  async find (params?: Params): Promise<Paginated<any> | any> {
+    return [{
       host: this.config.host,
       port: this.config.port,
       version: this.config.version
-    }
+    }]
   }
 }
 
-export default AuthClass
+export default StatusClass