+++ playground/hello/src/services/users/users.class.ts
@@ -1,22 +1,12 @@
-/* eslint-disable @typescript-eslint/no-empty-interface */
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import { NullableId, Params } from 'zapnode'
+import { Paginated, Params } from 'zapnode'
 import { MongoDBService } from 'zapnode-data'
-import { User, UserData } from './users.schema'
+import { User, UserData, UserQuery } from './users.schema'
 
-interface Query {
-}
-
-interface UsersParams extends Params<Query> {}
-
-class UsersClass extends MongoDBService<User> {
-  async find (params?: UsersParams) {
-    console.log('q=', params?.query)
-    return this.$find(params)
-  }
-
-  async create (data: UserData, params?: UsersParams) {
-    return this.$create(data, params)
+class UsersClass extends MongoDBService<User, UserData, Params<UserQuery>> {
+  async find (params?: Params<UserQuery>): Promise<User[]>
+  async find (params?: Params<UserQuery> & { query: { $paginate: true } }): Promise<Paginated<User>>
+  async find (params?: Params<UserQuery>): Promise<Paginated<User> | User[]> {
+    return this.repository.find(params)
   }
 }